requirements.txt
python-whois==0.6.5
requests==2.18.4
